requirements.txt
# Django
Django>=3.0.2,<3.1.0

# Django REST Framework
djangorestframework>=3.11.0,<3.12.0